Uses parent data, not progeny data to calculate allele freqs (see full commit message)
The method to estimate allele frequency in inferParents has changed to use parent data.
This is because this lead to issues in simulation data where one mom was used with random fathers.
There's a strong case to be made for using progeny genotypes to estimate allele frequency, but there's
more risk: one highly-mating father could bias freq estimate. In sim data, this lead to everything looking
like selfing.